A few artifact changes.
Undead Slayers get the Holy Spear of Light as a sacrifice gift now, as
was intended (going by the SLASH 6 Source Code).

Geeks get Shocker as their sacrifice gift, and SPFX_ATTK was added
to make its shock damage work correctly.

Jedi get Deluder, which used to be the second sacrifice gift for
Wizards.  Deluder was almost never used as their starting Cloak of Magic
Resistance was much better, and the tradeoffs were never worth it.

Similar to the above, Warriors get Deathsword instead of Barbarians.
Cleaver is enough for them for the time being.