Implement new summon invokes.
Both of these will be used in new quest artifacts for Acid and Electric
Mages.  There is a 1 in 4 chance of summoning the greater variant of
their respective elementals.

Both of the older summon invokes have been updated to have a chance of summoning
the greater versions of their elementals as well.