Fully implement the Forge.
Bluescreenofdeath added the Forge level from Nethack - The Next
Generation as well as the blacksmith.  However, the artifact creation
code was never implemented, making the level a bit of filler that had no
purpose.

This branch and commit is aiming to restore the blacksmith branch, as
well as expand its list of available artifacts.