Massive buffs to Blacksmith.
This is mostly to deter anything attempting to kill the blacksmith. This
also creates a new artifact specific to the blacksmith for self defense:

Warforger: Unaligned Heavy Hammer, +15 to-hit and +15 damage, chance
of doing an extra 6d6 damage and a chance to burn potions, scrolls, and
spellbooks.  Cannot be wished for.