Lunatic init changes
I intend to convert the Lunatic role into more of a homage to the
Lycanthrope role from SLASH 6.  To start, I've modified their initial
inventory.  Now it only consists of a orcish short sword, a potion of
sickness, and a random scroll.  Lunatics also start with an extra square
of vision similar to cavemen.