Slings can instakill Giants at Expert.
Adding in the rest of the Dragontamer patch.  This also includes a few
changes to the Sling of David.  It no longer does double damage against
only giants, it now applies its double damage against all creatures.
The Sling of David will also always instakill giants, regardless of the
users skill with slings.