A few item additions and changes.
This commit adds (or rather, revives) some items from previous variants.

*Sharpened Pencil: A wooden dagger that has low cost, does d2 damage to
small and large monsters, and can engrave (think of it as writing) as an
athame.  Comes from Nethack--.

*Expensive Suit: A cloak(as it was in its source) that grants MC2 and a
Charisma Bonus of 2.  Also comes from Nethack--

*Fire Helmet: An iron helmet that has an AC of 1 and grants Fire
Resistance.  Revived from SLASH 6.

*Ugly Backpack: It's literally a rename and color change of a sack.
Because of this it doesn't generate in the dungeon at all.  Again comes
from Nethack--.

Because of the additions there have been some changes when it comes to
some item probability:

Elven Dagger: 6->5
Elven Cloak: 8->6
Helmet: 8->6

There have also been a few changes when it comes to initial inventory
for a few roles.

*Undertakers start with an Expensive Suit instead of Victorian Underwear
(As they did in Nethack--).

*Firefighters gain a Fire Helmet and a Cloak of Quenching (only for the
AC bonus, not for the fire resistance as they already have this
intrensic).

*Graduates start with a +3 Sharpened Pencil instead of a Spoon, and an
Ugly Backpack instead of a Sack (Again, as they did in Nethack--).

The last change that should be listed is that the Goffic Backpack
artifact is now an ugly backpack instead of a sack.