Introduce a new level to the town branch.
Pleasantville: A mostly peaceful level with a hideout at the bottom
left corner of the map filled with thieves that contains the staircase
to the Town.  Also contains some random loot with a small chance of some
special easter eggs.

Level design is from the Lethe Patch.  This level might still need some
work, but it should be okay for now.