Revert Gauntlets of Power to use Slash'EM Behavior
Slash'EM Extended changed the Gauntlets of Power to give its strength
bonus only when enchanted, and even then it only incremented the
strength value by 1 (Str 18 -> Str 18/01). In this case, this made it
impossible for characters to throw Mjollnir.  This reverts that change,
allowing the Gauntlets of Power to function the same way they do in
Slash'EM.