A few minor edits.
Add colons to the bottom status bar, and make the lines a bit more
compatible with bot2_abbrev.  This also hides mov because for the most
part it's pointless, as it only changes during polymorph of if you're
Fast/Slow.

This also removes the "fixes" from BarclayII and AmyBSOD, as they don't
fix the problem at hand.  As such, we're back to segfaulting when too
many things are on screen.