Normalize the Gnomish Mines.
Previously, AmyBSOD changed the gnomish mines to generate random G and h
class enemies, which often produced monsters out of level or monsters
unrelated to the mines at all.  This commit reverts most enemy
placements in the mines back to vanilla, which should make the mines
much safer to traverse.