Completely remove the Activistor and Goff roles.
These two just were not clicking for most, and seemed a bit _too_ silly
for a roguelike.  Not borderline offensive as the previous two, but had
no real place here.  The activistor in particular was neutered heavily
after removing the topmodels and transvestites, while the goff were an
inferior vampire that gained reflection at level 10.  Twilight
reference?