Completely remove the shit trap.
Part 5 of "The Great Cleanup", which is mostly some minor things and
fixing anything screwed up/missing from the previous 4 parts.

Interesting concept, terrible terrible execution.  It's only seriously
used in the Spacewars fighter quest which is (eventually) going to get
an entire remake to be more accurate to Castle of the Winds anyway.
Other than that, it's just stupid.