Revert technique timeouts to rn1.
First step of a long series to remove most, if not all of the added rnz
calls from Slash'EM Extended.  The distribution for rnz is incredibly
strange, and really, really shouldn't be used for anything beyond prayer
timeout and artifact invokes.

Anyway, the only change for a player from this commit is that for the
most part technique timeouts should be lower on average.