Implement the Whips patch.
Adds the Chainwhip, Long Bullwhip, Thorned Chainwhip, and Long Thorned
Chainwhip to the game.  Chainwhips are not randomly generated in the
dungeon: they require a non-restricted whip user to wield an iron chain.
The Long Bullwhip and Long Thorned Chainwhip can be applied to a
location two squares away compared to one for the other whips, and
function similarly to a grappling hook.

This commit does NOT add the artifact chainwhip from the patch.