Add The Beastmaster's Duster from dNethack.
It's now the sacrifice gift for Zookeepers.  Liontamer is now not
associated with any role.  I figured it was a tad too useless for a
sacrifice gift for them.

The only changes between here and dNethack is that it's neutral aligned
and it grants an additional +2 AC.  The invoke is the same as it is in
dNethack.