Fully resolve the raise zombies human segfault
Turns out this was all due to the fact that the Human Zombie and Human
Mummy had MRACE_HUMAN.  In vanilla, human zombies and mummies don't have
an M2_HUMAN flag at all.  Due to having the MRACE_HUMAN flag, tamedog
nulls out the monster (as humans cannot be tamed) and the game
completely segfaults attempting to read the name of the now-nonexistant
monster.

This (and the previous commit) closes #29 on Github.