Revert Hand of Vecna attributes to Slash'EM
SLEX previously nerfed the Hand of Vecna by making it's Half Physical
Damage and Regeneration traits only activate if the player is weilding
the item.  This reverts to the way it's handled in Slash'EM; Half
Physical Damage and HP Regen are active as long as the Hand of Vecna is
carried.  Since the quest is more difficult the nerf seems even more
pointless.