Fix lightsource creation for the artifact lightsources
If you received one as a sacrifice gift, begin_burn would fail to locate
the object.

This moves starting the eternal lightsources to after the object has
been placed, whether in inventory or on the ground. This covers the
cases where the object is received from sacrifice gifts, started out in
a special level (flame mage quest for example), or was wished for.

There could be other ways to get one of these artifact lightsources, but
I couldn't think of any offhand.